fix packet build problem

+++ F767_MIRDC_TxNode/Core/Inc/packet.h
@@ -7,6 +7,6 @@
 #define PACKET_PAYLOAD_LEN  24
 #define PACKET_TOTAL_LEN    38
 
-uint16_t Packet_Build(uint8_t *packet , uint8_t freq_pair , uint8_t seq_id , uint8_t timestamp , IMU_Avg_t avg);
+uint16_t Packet_Build(uint8_t *packet , uint8_t freq_pair , uint8_t seq_id , uint64_t timestamp , IMU_Avg_t avg);
 
 #endif
